itest: move channel force close test into one file
